+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### Features
 
 - Better error messaging on parsing template errors
-- Infra: basic support of RDS: listing, creation and deletion of databases:  `awless list databases`; `awless create/delete database`
+- Infra: basic support of RDS: listing, creation and deletion of databases and database subnets:  `awless list databases`; `awless create/delete database/dbsubnet`
+- Infra: attach/detach an `instance` to a `targetgroup`
 - Access: create an AWS access key for a user
 - DNS: allow to revert creation/deletion of records
 - [#80](https://github.com/wallix/awless/issues/80) DNS: return the ChangeInfo id when creating/deleting a record